apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
pgvector
pyyaml
openai
sentence-transformers
pdfplumber
python-docx
readability-lxml
scikit-learn
